+++ springapp/src/main/java/com/example/springapp/repository/EnrollmentRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.springapp.model.Enrollment;
 
-public interface EnrollmentRepository extends JpaRepository<Enrollment,Integer>{
+public interface EnrollmentRepository extends JpaRepository<Enrollment,Long>{
 	
 }
